requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4
cffi==1.11.5
Click==7.0
cryptography==2.3.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.3.2
gpiozero==1.4.1
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.4.2
pylint==2.1.1
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
wrapt==1.10.11
